save politicians and test

+++ src/main/java/com/example/demo/dto/PoliticianDTO.java
@@ -1,6 +1,6 @@
 package com.example.demo.dto;
 
 @SuppressWarnings("preview")
-public record PoliticianDTO(String name, String id, Float rating) {
+public record PoliticianDTO(String name, String id, Double rating) {
 
 }
